Split the main file into separate small categorized files
