+++ src/misc.py
@@ -1,0 +1,26 @@
+from enum import Enum
+
+
+class SpeakerQuestion:
+    affiliation = "Company / Organization / Educational Institution"
+    homepage = "Social (Homepage)"
+    twitter = "Social (X/Twitter)"
+    mastodon = "Social (Mastodon)"
+    linkedin = "Social (LinkedIn)"
+    gitx = "Social (Github/Gitlab)"
+
+
+class SubmissionQuestion:
+    outline = "Outline"
+    tweet = "Abstract as a tweet / toot"
+    delivery = "My presentation can be delivered"
+    level = "Expected audience expertise"
+
+
+class SubmissionState(Enum):
+    accepted = "accepted"
+    confirmed = "confirmed"
+    withdrawn = "withdrawn"
+    rejected = "rejected"
+    canceled = "canceled"
+    submitted = "submitted"
